refactor: interface, constant, util 분리
