+++ client/src/components/common/StopWatch.tsx
@@ -1,23 +1,6 @@
-interface StopWatchProps {
-  elapsedSeconds: number;
-  isAnimationOn: boolean;
-}
+import type { StopWatch as StopWatchProps } from '@customTypes/Common';
 
-const secondsToHMS = (elapsedSeconds: number) => {
-  const hours = Math.floor(elapsedSeconds / 3600);
-  const minutes = Math.floor((elapsedSeconds % 3600) / 60);
-  const seconds = elapsedSeconds % 60;
-
-  return { hours, minutes, seconds };
-};
-
-const formatTime = (hours: number, minutes: number, seconds: number) => {
-  const paddedHours = hours.toString().padStart(2, '0');
-  const paddedMinutes = minutes.toString().padStart(2, '0');
-  const paddedSeconds = seconds.toString().padStart(2, '0');
-
-  return `${paddedHours} : ${paddedMinutes} : ${paddedSeconds}`;
-};
+import { secondsToHMS, formatTime } from '@utils/time';
 
 const StopWatch = ({ elapsedSeconds = 0, isAnimationOn }: StopWatchProps) => {
   const { hours, minutes, seconds } = secondsToHMS(elapsedSeconds);
