+++ client/src/constants/API.ts
@@ -1,0 +1,3 @@
+const API_BASE_URL = import.meta.env.DEV ? '/api' : import.meta.env.VITE_SIGNALING_SERVER_URL;
+
+export { API_BASE_URL };
