+++ client/src/constants/EMOTICONS.ts
@@ -1,0 +1,10 @@
+const SAD_EMOTICONS = [
+  '｡° ૮₍°´ᯅ`°₎ა °',
+  '｡° (ꢳࡇꢳ) °｡',
+  '༼ ༎ຶ ෴ ༎ຶ༽',
+  '꒰ 𖦹ˊᯅˋ𖦹 ꒱',
+  '✘ᴗ✘',
+  'ヽ(●´Д｀●)ﾉﾟ',
+];
+
+export { SAD_EMOTICONS };
