+++ client/src/constants/VIDEO.ts
@@ -1,0 +1,6 @@
+const RATIO = 4 / 3;
+const MAX_HEIGHT = 600;
+const MAX_WIDTH = MAX_HEIGHT * RATIO;
+const GAP = 8;
+
+export { MAX_HEIGHT, MAX_WIDTH, GAP };
