+++ client/src/hooks/useWebRTC.ts
@@ -1,36 +1,10 @@
 import { useState, useEffect, useRef } from 'react';
 import { Socket } from 'socket.io-client';
+
+import type { WebRTCData, WebRTCState, WebRTCControls } from '@customTypes/WebRTC';
+import { API_BASE_URL } from '@constants/API';
+
 import signalingClient from '@socket/signalingClient';
-
-interface WebRTCData {
-  peerConnection: RTCPeerConnection;
-  remoteStream: MediaStream;
-  dataChannel: RTCDataChannel;
-  nickName: string;
-}
-
-interface WebRTCState {
-  localStream: MediaStream;
-  webRTCMap: React.MutableRefObject<Map<string, WebRTCData>>;
-  participantCount: number;
-  grid: { cols: number; rows: number };
-  selectedVideoDeviceId: string;
-  selectedAudioDeviceId: string;
-}
-
-interface WebRTCControls {
-  toggleVideo: () => Promise<boolean>;
-  toggleMic: () => boolean;
-  joinRoom: (roomId: string) => Promise<Socket>;
-  exitRoom: () => void;
-  sendMessage: (message: string) => void;
-  getVideoDevices: () => Promise<MediaDeviceInfo[]>;
-  getAudioDevices: () => Promise<MediaDeviceInfo[]>;
-  setSelectedVideoDeviceId: React.Dispatch<React.SetStateAction<string>>;
-  setSelectedAudioDeviceId: React.Dispatch<React.SetStateAction<string>>;
-}
-
-const API_BASE_URL = import.meta.env.DEV ? '/api' : import.meta.env.VITE_SIGNALING_SERVER_URL;
 
 const useWebRTC = (): [WebRTCState, WebRTCControls] => {
   const socket = useRef<Socket>();
